# Support for close combinator for indexed effects