--- a/src/syntax/FStar.Syntax.Visit.fst
+++ b/src/syntax/FStar.Syntax.Visit.fst
@@ -282,11 +282,12 @@ let on_sub_wp_eff_combinators vfs (wpcs : wp_eff_combinators) : wp_eff_combinato
 let on_sub_layered_eff_combinators vfs (lecs : layered_eff_combinators) : layered_eff_combinators =
   let f_tscheme = f_tscheme vfs in
   {
-    l_repr         = on_tuple2 f_tscheme f_tscheme    lecs.l_repr;
-    l_return       = on_tuple2 f_tscheme f_tscheme    lecs.l_return;
-    l_bind         = on_tuple3 f_tscheme f_tscheme id lecs.l_bind;
-    l_subcomp      = on_tuple3 f_tscheme f_tscheme id lecs.l_subcomp;
-    l_if_then_else = on_tuple3 f_tscheme f_tscheme id lecs.l_if_then_else;
+    l_repr         = on_tuple2  f_tscheme f_tscheme           lecs.l_repr;
+    l_return       = on_tuple2  f_tscheme f_tscheme           lecs.l_return;
+    l_bind         = on_tuple3  f_tscheme f_tscheme id        lecs.l_bind;
+    l_subcomp      = on_tuple3  f_tscheme f_tscheme id        lecs.l_subcomp;
+    l_if_then_else = on_tuple3  f_tscheme f_tscheme id        lecs.l_if_then_else;
+    l_close        = map_option (on_tuple2 f_tscheme f_tscheme) lecs.l_close;
   }
 
 let on_sub_combinators vfs (cbs : eff_combinators) : eff_combinators =