… src/tosyntax/FStar.ToSyntax.ToSyntax.fst
… src/typechecker/FStar.TypeChecker.TcEffect.fst
… src/typechecker/FStar.TypeChecker.TcTerm.fst
… src/typechecker/FStar.TypeChecker.Util.fst
… src/typechecker/FStar.TypeChecker.Util.fsti
… ulib/FStar.Tactics.Effect.fsti